requirements.txt
streamlit
transformers
Pillow
pytube
openai
python-dotenv
requests
tf-keras
tensorflow